pyproject.toml
[project]
name = "ollama-agent"
version = "0.1.0"
description = "Local Ollama + Langfuse tracing example (chat + embeddings)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "langfuse>=2.44.0",
  "openai>=1.53.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.9.2",
  "typer>=0.12.5",
  "rich>=13.9.2",
  "chromadb>=0.4.22",
]

[tool.uv]
dev-dependencies = []


